llama_aws_eks_vllm_grpc_ttft_radha/requirements.txt
requests==2.31.0
kubernetes==29.0.0
locust==2.30.0

grpcio==1.65.5
grpcio-tools==1.65.5
grpcio-reflection==1.65.5
grpcio-health-checking==1.65.5
